refactor: edit MyPageMenuBar
- InkWell 위젯을 사용해 버튼으로 변경하였습니다.
- UiConfig 를 적용하였습니다.
